prevent del of x3mRouting.sh link if v2

+++ Install_x3mRouting.sh
@@ -5,13 +5,15 @@ while true; do
   /usr/sbin/curl -s --retry 3 https://raw.githubusercontent.com/Xentrk/x3mRouting/x3mRouting-NG/x3mRouting_Menu.sh -o /jffs/addons/x3mRouting/x3mRouting_Menu.sh
   chmod 755 /jffs/addons/x3mRouting/x3mRouting_Menu.sh
   # Remove old link or alias if exist
-  if [ -d "/opt/bin" ] && [ "$(/opt/bin/find /opt/bin/ -maxdepth 1 -type l -ls | grep -c "/opt/bin/x3mRouting -> /jffs/addons/x3mRouting/x3mRouting_Menu.sh")" -eq 0 ]; then
-    if [ -s "/opt/bin/x3mRouting" ]; then # old alias situation
-      rm -rf /opt/bin/x3mRouting 2>/dev/null
-    fi
-  else
+  if [ "$(/opt/bin/find /opt/bin/ -maxdepth 1 -type l -ls | grep -c "/opt/bin/x3mRouting -> /jffs/scripts/x3mRouting/x3mRouting.sh")" -eq 0 ]; then
+    if [ -d "/opt/bin" ] && [ "$(/opt/bin/find /opt/bin/ -maxdepth 1 -type l -ls | grep -c "/opt/bin/x3mRouting -> /jffs/addons/x3mRouting/x3mRouting_Menu.sh")" -eq 0 ]; then
+      if [ -s "/opt/bin/x3mRouting" ]; then # old alias situation
+        rm -rf /opt/bin/x3mRouting 2>/dev/null
+      fi
+    else
      rm -rf /opt/bin/x3mRouting 2>/dev/null
-  fi
+   fi
+ fi
   rm -rf /opt/bin/x3mMenu 2>/dev/null
   ln -s "/jffs/addons/x3mRouting/x3mRouting_Menu.sh" "/opt/bin/x3mMenu"
   x3mMenu
